On windows backslashes where created with topping exporter file paths. This leaded to issues when the relative path has been attached to an url.
The IliDataCache.download_file now:
- replaces the backslashes with slashes on URL
- normalizes the path to the current system

Still ili2db does not that.

The IliToppingMaker creating those paths is now storing everything with backslashes as well, nerveless if it's an url or a file path (because it cannot know that).

Windows can handel slashes in the path as well AFAIK, what means it's more stable to have slashes.